Create send_email.py
This python script creates a new folder in a location and sends the complete path of the folder via email using an exchange server. The name of the created folder is the date itself. This script is used by me on a daily schedule to send the location of the folder that contains various reports for the Finance department. The reports are loaded into this folder every day via a different cron job.